requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
pytest
apache-airflow
streamlit
plotly
